requirements.txt
future~=0.18.2
Pillow==9.4.0
jsonpickle~=3.0.1
tkinterdnd2~=0.3.0
pyglet~=2.0.3